Finished player tab and added auto commit to real transaction functions.

+++ src/main/java/com/mif/crew/premierleaguepostgres/DatabaseViewController.java
@@ -26,6 +26,8 @@ import java.io.IOException;
 import java.sql.*;
 import java.time.LocalDate;
 
+import static helper.DatabaseLogger.log;
+
 public class DatabaseViewController {
 
     private StandingsService standingsService;
@@ -59,6 +61,9 @@ public class DatabaseViewController {
     private Button changeMatchButton;
 
     @FXML
+    private Button clearFiltersButton;
+
+    @FXML
     private Button deleteMatchButton;
 
     @FXML
@@ -66,6 +71,9 @@ public class DatabaseViewController {
 
     @FXML
     private Button deleteTeamButton;
+
+    @FXML
+    private Button deletePlayerButton;
 
     @FXML
     private TextField homeGoalsText;
@@ -107,6 +115,15 @@ public class DatabaseViewController {
     private ListView<IdNamePair> playerList;
 
     @FXML
+    private TextField playerNameFilter;
+
+    @FXML
+    private ComboBox<String> playerPositionCombo;
+
+    @FXML
+    private ComboBox<String> playerTeamCombo;
+
+    @FXML
     private VBox playersPanel;
 
     @FXML
@@ -125,7 +142,7 @@ public class DatabaseViewController {
     private Button signPlayerButton;
 
     @FXML
-    private VBox signPlayerPanel;
+    private VBox filterPlayerPanel;
 
     @FXML
     private VBox squadPanel;
@@ -146,22 +163,32 @@ public class DatabaseViewController {
     private VBox createMatchPanel;
 
 
-
-
     private Connection conn;
 
 
 
     public void initialize() {
         DatabaseLogger.bind(infoArea);
-//        seasonListView.getSelectionModel().setSelectionMode(SelectionMode.MULTIPLE);
-//        playerList.getSelectionModel().setSelectionMode(SelectionMode.MULTIPLE);
+        DataHandler.getInstance().setMainController(this);
+
+        ObservableList<String> playerPositions = FXCollections.observableArrayList("All", "GK", "DF", "MF", "FW");
+        playerPositionCombo.setItems(playerPositions);
         applySeasonButton.setOnAction(e -> {
             Integer selected = seasonListView.getSelectionModel().getSelectedItem();
             if (selected != null) {
-                DataHandler.getInstance().setSelectedSeason(seasonListView.getSelectionModel().getSelectedItem());
+                DataHandler.getInstance().setSelectedSeason(selected);
+                deleteTeamButton.setDisable(true);
+                deleteMatchButton.setDisable(true);
+                changeMatchButton.setDisable(true);
+                deletePlayerButton.setDisable(true);
+                addTeamButton.setDisable(false);
+
+                squadPanel.setDisable(false);
                 createMatchPanel.setDisable(false);
+                filterPlayerPanel.setDisable(false);
                 refreshCurrentTab();
+                fillTeams(playerTeamCombo, selected);
+                playerTeamCombo.getItems().addFirst("All Teams");
                 fillTeams(homeTeamCombo, selected);
                 fillTeams(awayTeamCombo,selected);
                 fillWeeks(matchWeekCombo, selected);
@@ -215,10 +242,10 @@ public class DatabaseViewController {
                     if(QueryExecutor.runDDLQuery(conn, query, Integer.parseInt(newSeasonText.getText())))
                         seasonListView.getItems().add(Integer.parseInt(newSeasonText.getText()));
                 }else {
-                    DatabaseLogger.log("Season should be after the year 1992(the year premier league was founded");
+                    log("Season should be after the year 1992(the year premier league was founded");
                 }
             }catch(NumberFormatException ex){
-                DatabaseLogger.log("Season should be a number");
+                log("Season should be a number");
             }
         });
 
@@ -247,7 +274,14 @@ public class DatabaseViewController {
                         .and(matchWeekCombo.valueProperty().isNotNull())
                         .and(matchDatePicker.valueProperty().isNotNull())
                         .and(homeGoalsText.textProperty().isNotEmpty())
-                        .and(awayGoalsText.textProperty().isNotEmpty());
+                        .and(awayGoalsText.textProperty().isNotEmpty())
+                        .and(homeTeamCombo.valueProperty().isNotEqualTo(awayTeamCombo.valueProperty().get()));
+
+        clearFiltersButton.setOnAction(e -> {
+            playerTeamCombo.getSelectionModel().clearSelection();
+            playerPositionCombo.getSelectionModel().clearSelection();
+            playerNameFilter.clear();
+        });
 
         registerMatchButton.disableProperty().bind(allFieldsFilled.not());
 
@@ -257,6 +291,8 @@ public class DatabaseViewController {
 
         deleteTeamButton.setOnAction(this::deleteTeam);
 
+        deletePlayerButton.setOnAction(this::deletePlayer);
+
 
         addTeamButton.setOnAction(e -> popupHandle("TEAM"));
         addManagerButton.setOnAction(e -> popupHandle("MANAGER"));
@@ -265,12 +301,26 @@ public class DatabaseViewController {
 
     }
 
+    private void deletePlayer(ActionEvent actionEvent) {
+        ObservableList<String> selected = DataHandler.getInstance().getSelectedPlayer();
+        if (selected != null) {
+            String deletePositions = "DELETE FROM premierleague.playerposition WHERE playerid = ?";
+            if(QueryExecutor.runDDLQuery(conn, deletePositions, Integer.parseInt(selected.getFirst()))){
+                String query = "DELETE FROM premierleague.player WHERE playerid = ?";
+                if (QueryExecutor.runDDLQuery(conn, query, Integer.parseInt(selected.getFirst())))
+                    DataHandler.getInstance().getPlayersTable().getItems().removeIf(row -> row.getFirst().equals(selected.getFirst()));
+            }
+        }
+    }
+
     private void registerMatch(ActionEvent actionEvent) {
-        String query = "INSERT INTO PremierLeague.matches (home_team_Name, away_team_name, week, match_date, home_goals, away_goals, Season_End_Year) VALUES (?, ?, ?, ?, ?, ?, ?)";
+        String query = "INSERT INTO PremierLeague.matches (home_team_Name, away_team_name, week, match_date, home_goals, away_goals, Season_End_Year)" +
+                " VALUES (?, ?, ?, ?, ?, ?, ?)";
         Connection conn = DataHandler.getInstance().getConnection();
         int season = DataHandler.getInstance().getSelectedSeason();
 
-
+        int homeGoals = Integer.parseInt(homeGoalsText.getText());
+        int awayGoals = Integer.parseInt(awayGoalsText.getText());
 
         try (PreparedStatement stmt = conn.prepareStatement(query, Statement.RETURN_GENERATED_KEYS)) {
             stmt.setString(1, homeTeamCombo.getValue());
@@ -281,46 +331,53 @@ public class DatabaseViewController {
             stmt.setInt(6, Integer.parseInt(awayGoalsText.getText()));
             stmt.setInt(7, season);
 
-            ResultSet generatedKeys = stmt.getGeneratedKeys();
+            int affected = stmt.executeUpdate();
+
             Integer matchId = null;
+            ResultSet generatedKeys = stmt.getGeneratedKeys();
             if (generatedKeys.next()) {
                 matchId = generatedKeys.getInt(1);
-            }
 
-            int affected = stmt.executeUpdate();
+            }
 
             if (affected > 0) {
                 ObservableList<String> row = FXCollections.observableArrayList();
                 if(matchId == null){
-                    DatabaseLogger.log("Couldn't generate a match id for the new match");
+                    log("Couldn't generate a match id for the new match");
+                    return;
                 }
+                char result;
+                if(homeGoals > awayGoals) result = 'H';
+                else if(homeGoals < awayGoals) result = 'A';
+                else result = 'D';
                 row.add(String.valueOf(matchId));
-                row.add(homeTeamCombo.getValue());
-                row.add(awayTeamCombo.getValue());
                 row.add(String.valueOf(matchWeekCombo.getValue()));
                 row.add(String.valueOf(matchDatePicker.getValue()));
+                row.add(homeTeamCombo.getValue());
                 row.add(homeGoalsText.getText());
                 row.add(awayGoalsText.getText());
-                row.add(String.valueOf(season));
+                row.add(awayTeamCombo.getValue());
+                row.add(String.valueOf(result));
+
 
                 DataHandler.getInstance().getMatchesTable().getItems().add(row);
-                DatabaseLogger.log("Match added successfully");
+                log("Match added successfully");
             }
         } catch (SQLException e) {
-            DatabaseLogger.log("Failed to insert into TeamSeason: " + e.getMessage());
+            log("Failed to insert into TeamSeason: " + e.getMessage());
         }
     }
 
     private void fillWeeks(ComboBox<Integer> matchWeekCombo, Integer selected) {
         if (conn == null) {
-            DatabaseLogger.log("No database connection.");
+            log("No database connection.");
             return;
         }
 
         String query = "SELECT count(distinct team_name) FROM premierleague.teamseason " +
                 "WHERE season_end_year = ?";
 
-        DatabaseLogger.log("Running query: " + query);
+        log("Running query: " + query);
 
         try (PreparedStatement stmt = conn.prepareStatement(query)) {
             stmt.setInt(1, selected);
@@ -333,7 +390,7 @@ public class DatabaseViewController {
                 }
             }
         } catch (SQLException e) {
-            DatabaseLogger.log("Failed to load data: " + e.getMessage());
+            log("Failed to load data: " + e.getMessage());
             matchesPanel.getChildren().add(new Label("Failed to load data."));
         }
 
@@ -341,14 +398,14 @@ public class DatabaseViewController {
 
     private void fillTeams(ComboBox<String> teamCombo, Integer selected) {
         if (conn == null) {
-            DatabaseLogger.log("No database connection.");
+            log("No database connection.");
             return;
         }
 
-        String query = "SELECT team_name FROM premierleague.team " +
+        String query = "SELECT team_name FROM premierleague.teamseason " +
                 "WHERE season_end_year = ?";
 
-        DatabaseLogger.log("Running query: " + query);
+        log("Running query: " + query);
 
         try (PreparedStatement stmt = conn.prepareStatement(query)) {
             stmt.setInt(1, selected);
@@ -361,14 +418,14 @@ public class DatabaseViewController {
                 teamCombo.setItems(teams);
             }
         } catch (SQLException e) {
-            DatabaseLogger.log("Failed to load data: " + e.getMessage());
+            log("Failed to load data: " + e.getMessage());
             matchesPanel.getChildren().add(new Label("Failed to load data."));
         }
     }
 
     private void deleteMatch(ActionEvent actionEvent) {
-        TableView<ObservableList<String>> standingsTable = DataHandler.getInstance().getStandingsTable();
-        ObservableList<String> selectedRow = standingsTable.getSelectionModel().getSelectedItem();
+        TableView<ObservableList<String>> matchesTable = DataHandler.getInstance().getMatchesTable();
+        ObservableList<String> selectedRow = matchesTable.getSelectionModel().getSelectedItem();
 
         int matchId = Integer.parseInt(selectedRow.getFirst());
 
@@ -379,39 +436,53 @@ public class DatabaseViewController {
 
             int affected = stmt.executeUpdate();
             if (affected > 0) {
-                standingsTable.getItems().remove(selectedRow);
-                DatabaseLogger.log("Deleted match" + matchId);
+                matchesTable.getItems().remove(selectedRow);
+                log("Deleted match" + matchId);
             } else {
-                DatabaseLogger.log("Team not found or already deleted.");
+                log("Team not found or already deleted.");
             }
         } catch (SQLException e) {
-            DatabaseLogger.log("Error deleting team from TeamSeason: " + e.getMessage());
+            log("Error deleting team from TeamSeason: " + e.getMessage());
         }
+
     }
 
     private void deleteTeam(ActionEvent actionEvent) {
         TableView<ObservableList<String>> standingsTable = DataHandler.getInstance().getStandingsTable();
-        ObservableList<String> selectedRow = standingsTable.getSelectionModel().getSelectedItem();
+        ObservableList<String> selected = standingsTable.getSelectionModel().getSelectedItem();
 
-        String teamName = selectedRow.get(1);
-        int season = Integer.parseInt(selectedRow.get(0));
-
-        String query = "DELETE FROM PremierLeague.TeamSeason WHERE team_name = ? AND season_end_year = ?";
-
-        try (PreparedStatement stmt = DataHandler.getInstance().getConnection().prepareStatement(query)) {
-            stmt.setString(1, teamName);
-            stmt.setInt(2, season);
-
-            int affected = stmt.executeUpdate();
-            if (affected > 0) {
-                standingsTable.getItems().remove(selectedRow);
-                DatabaseLogger.log("Deleted team " + teamName + " from season " + season);
-            } else {
-                DatabaseLogger.log("Team not found or already deleted.");
+        try {
+            if (selected != null) {
+                conn.setAutoCommit(false);
+                int season = DataHandler.getInstance().getSelectedSeason();
+                String teamName = selected.getFirst();
+                String deleteTeam = "DELETE FROM PremierLeague.TeamSeason WHERE team_name = ? AND season_end_year = ?";
+                if (QueryExecutor.runDDLQuery(conn, deleteTeam, teamName, season)) {
+                    DatabaseLogger.log("Deleted team " + teamName);
+                    String deleteMatches = "DELETE FROM premierleague.matches WHERE home_team_name = ? OR away_team_name = ?";
+                    if (QueryExecutor.runDDLQuery(conn, deleteMatches, teamName, teamName)) {
+                        DatabaseLogger.log("Deleted matches associated with " + teamName);
+                        conn.commit();
+                        refreshCurrentTab();
+                    }
+                }
             }
         } catch (SQLException e) {
-            DatabaseLogger.log("Error deleting team from TeamSeason: " + e.getMessage());
+            DatabaseLogger.log(e.getMessage());
+            try {
+                conn.rollback();
+            } catch (SQLException ex) {
+                throw new RuntimeException(ex);
+            }
+        }finally{
+            try {
+                conn.setAutoCommit(true);
+            } catch (SQLException e) {
+                throw new RuntimeException(e);
+            }
         }
+
+
     }
 
     private void popupHandle(String name) {
@@ -433,7 +504,7 @@ public class DatabaseViewController {
                 stage.initModality(Modality.APPLICATION_MODAL);
                 stage.showAndWait();
             } catch (IOException e) {
-                DatabaseLogger.log(e.getMessage());
+                log(e.getMessage());
             }
         }
     }
@@ -450,7 +521,6 @@ public class DatabaseViewController {
             seasonsService = new SeasonService(conn);
             matchService = new MatchService(conn);
             playerService = new PlayerService(conn);
-
 
 
             seasonsService.loadSeasons(seasonListView);
@@ -490,7 +560,7 @@ public class DatabaseViewController {
     }
 
     public void loadStandingsView() {
-        Integer season = seasonListView.getSelectionModel().getSelectedItem();
+        Integer season = DataHandler.getInstance().getSelectedSeason();
 
         if (season == null) {
             standingsPanel.getChildren().setAll(new Label("Please select one of the seasons."));
@@ -500,7 +570,6 @@ public class DatabaseViewController {
         standingsService.loadStandingsTable(season, standingsPanel, squadService, squadTitle, playerList, managerList);
 
         TableView<ObservableList<String>> standingsTable = DataHandler.getInstance().getStandingsTable();
-        deleteTeamButton.setDisable(true);
         standingsTable.getSelectionModel().selectedItemProperty().addListener((obs, oldSel, newSel) -> {
             deleteTeamButton.setDisable(newSel == null);
         });
@@ -509,19 +578,19 @@ public class DatabaseViewController {
 
 
     public void loadMatchesView() {
-        Integer season = seasonListView.getSelectionModel().getSelectedItem();
+        Integer season = DataHandler.getInstance().getSelectedSeason();
 
         if (season == null) {
-            standingsPanel.getChildren().setAll(new Label("Please select one of the seasons."));
+            matchesPanel.getChildren().setAll(new Label("Please select one of the seasons."));
             return;
         }
 
         matchService.loadMatchTable(season, matchesPanel);
 
         TableView<ObservableList<String>> matchesTable = DataHandler.getInstance().getMatchesTable();
-        deleteMatchButton.setDisable(true);
         matchesTable.getSelectionModel().selectedItemProperty().addListener((obs, oldSel, newSel) -> {
             deleteMatchButton.setDisable(newSel == null);
+            changeMatchButton.setDisable(newSel == null);
         });
     }
 
@@ -536,23 +605,43 @@ public class DatabaseViewController {
                 return new Pair<>(start, end);
             }
         } catch (SQLException e) {
-            DatabaseLogger.log("Failed to fetch season date range: " + e.getMessage());
+            log("Failed to fetch season date range: " + e.getMessage());
         }
         return null;
     }
 
 
     public void loadPlayersView() {
-        squadPanel.getChildren().setAll(createPlayerDetailsPanel());
-        configureTableForPlayers();
+        Integer season = DataHandler.getInstance().getSelectedSeason();
+
+        if (season == null) {
+            playersPanel.getChildren().setAll(new Label("Please select one of the seasons."));
+            return;
+        }
+
+
+        updatePlayerTable(season);
+
+        playerTeamCombo.valueProperty().addListener((obs, oldVal, newVal) -> updatePlayerTable(season));
+        playerPositionCombo.valueProperty().addListener((obs, oldVal, newVal) -> updatePlayerTable(season));
+        playerNameFilter.textProperty().addListener((obs, oldVal, newVal) -> updatePlayerTable(season));
+
+        TableView<ObservableList<String>> playersTable = DataHandler.getInstance().getPlayersTable();
+        playersTable.getSelectionModel().selectedItemProperty().addListener((obs, oldSel, newSel) -> {
+            deletePlayerButton.setDisable(newSel == null);
+        });
     }
 
-    private void configureTableForPlayers() {
+    private void updatePlayerTable(int season) {
+        String selectedTeam = playerTeamCombo.getValue();
+        String selectedPosition = playerPositionCombo.getValue();
+        String nameFilter = playerNameFilter.getText();
 
-    }
+        if ("All Teams".equals(selectedTeam)) selectedTeam = null;
+        if ("All".equals(selectedPosition)) selectedPosition = null;
+        if (nameFilter != null && nameFilter.isBlank()) nameFilter = null;
 
-    private Node createPlayerDetailsPanel() {
-        return null;
+        playerService.loadPlayerTable(season, playersPanel, selectedTeam, selectedPosition, nameFilter);
     }
 
 
@@ -565,10 +654,6 @@ public class DatabaseViewController {
         } else if (selectedTab.getText().equalsIgnoreCase("PLAYERS")) {
             loadPlayersView();
         }
-    }
-
-    public void log(String message) {
-        infoArea.appendText(message + "\n");
     }
